[private-storage] Improved PR #20
- made private storage tests reusable
- updated migration number
- updated root urlconf example in README

Related to #20
