+++ .gitignore
@@ -47,7 +47,8 @@ coverage.xml
 # Django stuff:
 *.log
 tests/openwisp2/media/
-!tests/openwisp2/media/fake-img.bin
+tests/openwisp2/firmware/
+!tests/openwisp2/firmware/fake-img.bin
 
 # Sphinx documentation
 docs/_build/
