+++ tests/openwisp2/sample_firmware_upgrader/tests.py
@@ -6,6 +6,7 @@ from django.urls import reverse
 from openwisp_firmware_upgrader.tests.base import TestUpgraderMixin
 from openwisp_firmware_upgrader.tests.base.test_admin import BaseTestAdmin
 from openwisp_firmware_upgrader.tests.base.test_models import BaseTestModels, BaseTestModelsTransaction
+from openwisp_firmware_upgrader.tests.base.test_private_storage import BasePrivateStorage
 from swapper import load_model
 
 BatchUpgradeOperation = load_model('firmware_upgrader', 'BatchUpgradeOperation')
@@ -42,6 +43,17 @@ class TestModels(BaseTestModels, TestUpgraderMixin, TestCase):
 @skipUnless(os.environ.get('SAMPLE_APP', False),
             'Running tests on standard openwisp_firmware_upgrader models')
 class TestModelsTransaction(BaseTestModelsTransaction, TestUpgraderMixin, TransactionTestCase):
+    device_firmware_model = DeviceFirmware
+    upgrade_operation_model = UpgradeOperation
+    batch_upgrade_operation_model = BatchUpgradeOperation
+    firmware_image_model = FirmwareImage
+    build_model = Build
+    category_model = Category
+
+
+@skipUnless(os.environ.get('SAMPLE_APP', False),
+            'Running tests on standard openwisp_firmware_upgrader models')
+class TestPrivateStorage(BasePrivateStorage, TestUpgraderMixin, TestCase):
     device_firmware_model = DeviceFirmware
     upgrade_operation_model = UpgradeOperation
     batch_upgrade_operation_model = BatchUpgradeOperation
